(refactor) Rename business schema to contacts schema across all files
