+++ database-mongo/models/user.js
@@ -8,13 +8,13 @@ var userSchema = mongoose.Schema({
   name: String,
   password: String,
   requests: [{ type: mongoose.Schema.Types.ObjectId, ref: 'Requests'}],
+  facebook: Object
   // name: String,
   // userCellPhone: String,
   // userEmail: String,
   // username: { type: String, unique: true },
   // password: String,
   // userEmail: String,
-  facebook: Object,
   // userCellPhone: String,
   // salt: String,
   // userAddress: String,
