--- npm-debug.log
@@ -1,48 +1,0 @@
-0 info it worked if it ends with ok
-1 verbose cli [ '/usr/local/Cellar/node/7.7.4/bin/node',
-1 verbose cli   '/usr/local/bin/npm',
-1 verbose cli   'run',
-1 verbose cli   'server-dev' ]
-2 info using npm@4.1.2
-3 info using node@v7.7.4
-4 verbose run-script [ 'preserver-dev', 'server-dev', 'postserver-dev' ]
-5 info lifecycle github-fetcher-fullstack-v2@1.0.0~preserver-dev: github-fetcher-fullstack-v2@1.0.0
-6 silly lifecycle github-fetcher-fullstack-v2@1.0.0~preserver-dev: no script for preserver-dev, continuing
-7 info lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: github-fetcher-fullstack-v2@1.0.0
-8 verbose lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: unsafe-perm in lifecycle true
-9 verbose lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: PATH: /usr/local/lib/node_modules/npm/bin/node-gyp-bin:/Users/tinarko/Documents/HR/Legacy/Quoted/node_modules/.bin:/usr/local/bin:/usr/bin:/bin:/usr/sbin:/sbin
-10 verbose lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: CWD: /Users/tinarko/Documents/HR/Legacy/Quoted
-11 silly lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: Args: [ '-c', 'node server/index.js' ]
-12 silly lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: Returned: code: 1  signal: null
-13 info lifecycle github-fetcher-fullstack-v2@1.0.0~server-dev: Failed to exec server-dev script
-14 verbose stack Error: github-fetcher-fullstack-v2@1.0.0 server-dev: `node server/index.js`
-14 verbose stack Exit status 1
-14 verbose stack     at EventEmitter.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/lifecycle.js:279:16)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at EventEmitter.emit (events.js:194:7)
-14 verbose stack     at ChildProcess.<anonymous> (/usr/local/lib/node_modules/npm/lib/utils/spawn.js:40:14)
-14 verbose stack     at emitTwo (events.js:106:13)
-14 verbose stack     at ChildProcess.emit (events.js:194:7)
-14 verbose stack     at maybeClose (internal/child_process.js:899:16)
-14 verbose stack     at Process.ChildProcess._handle.onexit (internal/child_process.js:226:5)
-15 verbose pkgid github-fetcher-fullstack-v2@1.0.0
-16 verbose cwd /Users/tinarko/Documents/HR/Legacy/Quoted
-17 error Darwin 16.4.0
-18 error argv "/usr/local/Cellar/node/7.7.4/bin/node" "/usr/local/bin/npm" "run" "server-dev"
-19 error node v7.7.4
-20 error npm  v4.1.2
-21 error code ELIFECYCLE
-22 error github-fetcher-fullstack-v2@1.0.0 server-dev: `node server/index.js`
-22 error Exit status 1
-23 error Failed at the github-fetcher-fullstack-v2@1.0.0 server-dev script 'node server/index.js'.
-23 error Make sure you have the latest version of node.js and npm installed.
-23 error If you do, this is most likely a problem with the github-fetcher-fullstack-v2 package,
-23 error not with npm itself.
-23 error Tell the author that this fails on your system:
-23 error     node server/index.js
-23 error You can get information on how to open an issue for this project with:
-23 error     npm bugs github-fetcher-fullstack-v2
-23 error Or if that isn't available, you can get their info via:
-23 error     npm owner ls github-fetcher-fullstack-v2
-23 error There is likely additional logging output above.
-24 verbose exit [ 1, true ]
